revert : delete createBooking method

+++ src/main/java/com/tiko/tiko/Booking/DTOs/BookingItemRequest.java
@@ -8,6 +8,6 @@ public record BookingItemRequest(
         Long eventTicketPriceId,
 
         @Positive
-        int quantity
+        Long quantity
 ) {
 }
